--- src/components/Modal/index.ts
@@ -1,1 +1,0 @@
-export { default as Modal } from './Modal';
