+++ src/components/Popover/Actions.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
 import { type ReactNode } from 'react';
-import styles from './Modal.module.css';
+import styles from './Popover.module.css';
 
 interface Props {
   children: ReactNode;
