+++ src/components/Popover/index.ts
@@ -1,0 +1,1 @@
+export { default as Popover } from './Popover';
