+++ src/components/index.ts
@@ -3,7 +3,7 @@ export { CommentCard } from './CommentCard';
 export { CommentCardSkeleton } from './CommentCardSkeleton';
 export { CommentForm } from './CommentForm';
 export { CommentList } from './CommentList';
-export { Modal } from './Modal';
+export { Popover } from './Popover';
 export { PostCard } from './PostCard';
 export { PostCardSkeleton } from './PostCardSkeleton';
 export { PostForm } from './PostForm';
